requirements.txt
protobuf
transformers==4.30.2
cpm_kernels
torch>=2.0
gradio==3.39
mdtex2html
sentencepiece
accelerate
sse-starlette
streamlit>=1.24.0
fastapi
typing_extensions==4.4.0
uvicorn
sse_starlette
pydantic

python-dateutil
huggingface_hub
langchain
InstructorEmbedding
sentence_transformers
arrow
pytz
tiktoken
